File restructure for readability; docstrings additions
