--- config.local.ini
@@ -1,4 +1,0 @@
-[DEFAULT]
-database_path_root = /Users/andrewpurchase/Documents/electicity-demand/data/test_data
-app_path_root = /Users/andrewpurchase/Documents/electicity-demand
-
