+++ config/city_main.yaml
@@ -6,7 +6,7 @@ file_extract_arg_map:
   extract_write_options: write_options
   extract_write_mode: write_mode
   
-geonames_clean_arg_map:
+city_clean_arg_map:
   clean_input_table: input_table
   clean_output_table: output_table
   clean_write_options: write_options
