+++ pipeline/jobs/city/city_clean.py
@@ -1,7 +1,9 @@
-from src.utils.config_loader import get_env_variable, load_config_file
-from src.utils import shared_helpers as sh
+from pipeline.utils.config_loader import get_env_variable, load_config_file
+from pipeline.utils import shared_helpers as sh
+from pipeline.utils import config_loader as cl
 import argparse
 import json
+import pyspark.sql.functions as sf
 
 
 def main(
@@ -11,7 +13,7 @@ def main(
     write_mode: str,
     spark=None,
 ):
-    """Clean geonames data and save it as a table."""
+    """Clean city data and save it as a table."""
 
     # Convert options from string to dict if needed
     if isinstance(write_options, str):
@@ -19,19 +21,22 @@ def main(
 
     # Retrieve app root based on environment
     env = get_env_variable()
-    config = load_config_file(env)
-    app_root = config.get("app_path_root")
+    config_params = load_config_file(env)
+    app_root = config_params.get("app_path_root")
 
     # Defining column mappings
-    job_config_file = f"{app_root}/config/geonames_clean.yaml"
-    column_mappings, column_select = sh.load_column_config(job_config_file)
+    job_config_file = f"{app_root}/config/city_clean.yaml"
+    column_mappings, column_select = sh.load_config_keys(job_config_file, "column_mappings", "column_select")
 
     # Load the extracted data
     file_df = spark.read.format("delta").load(input_table)
 
     # Split coordinates column into latitude and longitude
-    updated_coords_df = sh.split_coordinates(file_df, "Coordinates")
-
+    updated_coords_df = (
+        file_df.withColumn("longitude", sf.substring_index(file_df["Coordinates"], ",", 1))
+        .withColumn("latitude", sf.substring_index(file_df["Coordinates"], ",", -1))
+        .drop("Coordinates")
+    )
     # Rename columns based on the mapping
     renamed_df = sh.rename_columns(updated_coords_df, column_mappings)
 
@@ -44,18 +49,18 @@ def main(
 
 if __name__ == "__main__":
     # Set up argument parser
-    parser = argparse.ArgumentParser(description="This script is used to clean geonames files")
+    parser = argparse.ArgumentParser()
 
     # Define expected arguments
     parser.add_argument("--input_table", help="", required=True)
     parser.add_argument("--output_table", help="", required=True)
     parser.add_argument("--write_options", help="", default='{"header": "true"}')
     parser.add_argument("--write_mode", help="", default="append")
-    parser.add_argument("--spark_app_name", help="", default="Geonames_Clean_Job")
+    parser.add_argument("--spark_app_name", help="", default="City_Clean_Job")
 
     # Parse the arguments
     args = parser.parse_args()
     args.write_options = json.loads(args.write_options)
 
     # Run main
-    sh.run_job(main, args)
+    cl.run_job(main, args)
