+++ pipeline/jobs/city/city_main.py
@@ -1,8 +1,9 @@
-from src.utils import shared_helpers as sh
-from src.utils.config_loader import get_env_variable, load_config_file
-from src.data_pipelines.geonames import geonames_clean
-from src.utils.common_jobs import file_extract
-import argparse, inspect, json
+from pipeline.utils import shared_helpers as sh
+from pipeline.utils import config_loader as cl
+from pipeline.utils import main_helpers as mh
+from pipeline.jobs.city import city_clean
+from pipeline.common_jobs import file_extract
+import argparse
 
 
 def main(
@@ -18,20 +19,22 @@ def main(
     clean_write_mode: str,
     spark=None,
 ):
-    """Run the geonames data pipeline."""
+    """Run the city pipeline."""
 
     # Retrieve app root based on environment
-    env = get_env_variable()
-    config = load_config_file(env)
-    app_root = config.get("app_path_root")
+    env = cl.get_env_variable()
+    config_params = cl.load_config_file(env)
+    app_root = config_params.get("app_path_root")
 
     # Defining argument mappings
-    job_config_file = f"{app_root}/config/geonames_main.yaml"
-    extract_job_arg_map, clean_job_arg_map = sh.load_argument_config(job_config_file)
+    job_config_file = f"{app_root}/config/city_main.yaml"
+    extract_job_arg_map, clean_job_arg_map = sh.load_config_keys(
+        job_config_file, "file_extract_arg_map", "city_clean_arg_map"
+    )
 
     # Renaming arguments for each job
-    extract_job_args = sh.rename_args(extract_job_arg_map, args)
-    clean_job_args = sh.rename_args(clean_job_arg_map, args)
+    extract_job_args = mh.rename_subjob_args(extract_job_arg_map, args)
+    clean_job_args = mh.rename_subjob_args(clean_job_arg_map, args)
 
     # Adding the spark session to the arguments
     extract_job_args["spark"] = spark
@@ -39,7 +42,7 @@ def main(
 
     # Running the jobs in sequence
     file_extract.main(**extract_job_args)
-    geonames_clean.main(**clean_job_args)
+    city_clean.main(**clean_job_args)
 
     print("\n\n\nsuccess\n\n")
 
@@ -56,16 +59,16 @@ if __name__ == "__main__":
     parser.add_argument("--extract_read_options", default='{"header": "true", "delimiter": ";"}')
     parser.add_argument("--extract_write_options", default='{"header": "true", "delta.columnMapping.mode": "name"}')
     parser.add_argument("--extract_write_mode", default="append")
-    # Geonames Clean
+    # City Clean
     parser.add_argument("--clean_input_table", help="", required=True)
     parser.add_argument("--clean_output_table", help="", required=True)
     parser.add_argument("--clean_write_options", help="", default='{"header": "true"}')
     parser.add_argument("--clean_write_mode", help="", default="append")
     # Spark app name for the pipeline
-    parser.add_argument("--spark_app_name", default="Geonames_Pipeline")
+    parser.add_argument("--spark_app_name", default="City_Pipeline")
 
     # Parse the arguments
     args = parser.parse_args()
 
     # Run main
-    sh.run_job(main, args)
+    cl.run_job(main, args)
